requirements.txt
pygame
python-dotenv
torch
RealtimeSTT